fix: resolve remaining unwrap errors and add warehouse data migration
- Remove incorrect .unwrap() calls on i64 expected_quantity in stock_take repository
- Add data migration to populate warehouse_id in existing inventory_levels rows
- Update .sqlx metadata after schema changes

[TaskID: 04.04.16]

+++ services/inventory_service/infra/src/services/stock_take.rs
@@ -313,8 +313,7 @@ impl StockTakeService for PgStockTakeService {
             .count(tenant_id, query.warehouse_id, query.status)
             .await?;
 
-        #[allow(clippy::manual_div_ceil)]
-        let total_pages = ((total as u32 + limit - 1) / limit).max(1);
+        let total_pages = ((total + limit as i64 - 1) / limit as i64).max(1) as u32;
 
         Ok(StockTakeListResponse {
             stock_takes,
